Rename `GameType` to `MatchType` and make `enum` instead of `class`
