+++ osu.Game/Screens/OnlinePlay/OnlinePlayComposite.cs
@@ -30,7 +30,7 @@ namespace osu.Game.Screens.OnlinePlay
         protected Bindable<RoomStatus> Status { get; private set; }
 
         [Resolved(typeof(Room))]
-        protected Bindable<GameType> Type { get; private set; }
+        protected Bindable<MatchType> Type { get; private set; }
 
         [Resolved(typeof(Room))]
         protected BindableList<PlaylistItem> Playlist { get; private set; }
